Estado del servicio de mapa de camas ahora se obtiene desde el servicio de la organizacion como un observable
