+++ src/app/components/mapa-de-camas/mapa-de-camas/mapa-de-camas.component.html
@@ -10,7 +10,7 @@
 
                     <hr />
 
-                    <div class="row">
+                    <div class="row" *ngIf="estadoServicio && estadoServicio.total > 0">
                         <div class="col-2">
                             <h4>
                                 Estado del servicio >
