+++ src/app/services/organizacion.service.ts
@@ -60,4 +60,66 @@ export class OrganizacionService {
         establecimiento.activo = true;
         return this.save(establecimiento);
     }
+
+    /**
+     * Devuelve el estado del servicio del mapa de cama consultado de una organización
+     * Si paso como parámetro el servicioHospitalario, filtrará las camas que pertenezcan a ese servicio.
+     *
+     * @param {any} camas Array de camas a obtener los resultados
+     * @param {any} [servicioHospitalario=null]
+     * @returns {any} Object {total, desocupadas, ocupadas, descontaminacion, reparacion, bloqueadas, desocupadasOxigeno}
+     * @memberof OrganizacionService
+     */
+    getEstadoServicio(camas, servicioHospitalario = null): Observable<any> {
+        return new Observable((observer) => {
+
+            // si paso el servicio entonces filtro primero por ese servicio
+            if (servicioHospitalario) {
+                camas = camas.filter(cama => cama.servicio.id === servicioHospitalario);
+            }
+
+            const ocupadas = camas.filter(function (i) {
+                return (i.ultimoEstado.estado === 'ocupada');
+            });
+
+            // ocupacion
+            const bloqueadas = camas.filter(function (i) {
+                return (i.ultimoEstado.estado === 'bloqueada');
+            });
+
+            // descontaminadas
+            const descontaminacion = camas.filter(function (i) {
+                return (i.ultimoEstado.estado === 'desocupada' && !i.desinfectada);
+            });
+
+            // reparación
+            const reparacion = camas.filter(function (i) {
+                return (i.ultimoEstado.estado === 'reparacion');
+            });
+
+            // disponibles
+            const desocupadas = camas.filter(function (i) {
+                return (i.ultimoEstado.estado === 'desocupada');
+            });
+
+            // desocupadas y con oxígeno - TODO: modificar filtro una vez que se defina
+            const desocupadasOxigeno = camas.filter(function (i) {
+                return (i.ultimoEstado.estado === 'desocupada' && i.oxigeno);
+            });
+
+            const estado = {
+                'total': camas.length,
+                'ocupadas': ocupadas.length,
+                'desocupadas': desocupadas.length,
+                'descontaminacion': descontaminacion.length,
+                'reparacion': reparacion.length,
+                'bloqueadas': bloqueadas.length,
+                'desocupadasOxigeno': desocupadasOxigeno.length
+            };
+
+            observer.next(estado);
+
+            observer.complete();
+        });
+    }
 }
